Make checking for existance in handlers consistent
While some handlers were already using this setup others were simply assigning the value to the variable that is later used for the put operation which an IDE could flag as an unnecessary assignment, although it is actually used to assert that the object already exists. This is not at all clear from the previous code but this way it is.